fix: extract preview helpers to _fleet_preview.py, use package-level RecipeKind import, fix mock lambda
- Move _build_dispatch_recipe_table and _print_dispatch_preview to new
  _fleet_preview.py to satisfy REQ-IMP-007 (no submodule imports) and
  CD6 (fleet/__init__.py ≤400 lines)
- Move render_fleet_error to _fleet_display.py, re-export from __init__
- Replace autoskillit.recipe.schema.RecipeKind with autoskillit.recipe.RecipeKind
- Update test_fleet_reload_relaunches_without_resume mock to accept recipe_table kwarg
