+++ src/autoskillit/cli/fleet/_fleet_display.py
@@ -268,3 +268,25 @@ def _watch_loop(state_path: Path) -> int:
             termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
         except (termios.error, OSError):
             pass
+
+
+def render_fleet_error(envelope_json: str) -> int:
+    """Render a fleet error envelope to stderr.
+
+    Returns exit code: 3 for fleet envelope errors, 0 for non-error envelopes.
+    """
+    import json
+
+    try:
+        data = json.loads(envelope_json)
+    except (json.JSONDecodeError, TypeError):
+        return 0
+    if data.get("success") is not False:
+        return 0
+    msg = data.get("user_visible_message") or "unknown error"
+    code = data.get("error", "")
+    sys.stderr.write(f"fleet error [{code}]: {msg}\n")
+    details = data.get("details")
+    if details:
+        sys.stderr.write(f"  details: {json.dumps(details)}\n")
+    return 3
